+++ fractal_server/app/schemas/v2/dataset.py
@@ -12,8 +12,9 @@ from pydantic.types import AwareDatetime
 
 from .._filter_validators import validate_attribute_filters
 from .._filter_validators import validate_type_filters
+from .._validators import cant_set_none
+from .._validators import NonEmptyString
 from .._validators import root_validate_dict_keys
-from .._validators import valstr
 from .project import ProjectReadV2
 from fractal_server.images import SingleImage
 from fractal_server.images.models import AttributeFiltersType
@@ -23,7 +24,7 @@ from fractal_server.urls import normalize_url
 class DatasetCreateV2(BaseModel):
     model_config = ConfigDict(extra="forbid")
 
-    name: str
+    name: NonEmptyString
 
     zarr_dir: Optional[str] = None
 
@@ -41,8 +42,6 @@ class DatasetCreateV2(BaseModel):
     _attribute_filters = field_validator("attribute_filters")(
         classmethod(validate_attribute_filters)
     )
-
-    _name = field_validator("name")(classmethod(valstr("name")))
 
     @field_validator("zarr_dir")
     @classmethod
@@ -73,7 +72,7 @@ class DatasetReadV2(BaseModel):
 class DatasetUpdateV2(BaseModel):
     model_config = ConfigDict(extra="forbid")
 
-    name: Optional[str] = None
+    name: Optional[NonEmptyString] = None
     zarr_dir: Optional[str] = None
     type_filters: Optional[dict[str, bool]] = None
     attribute_filters: Optional[dict[str, list[Any]]] = None
@@ -90,7 +89,10 @@ class DatasetUpdateV2(BaseModel):
         classmethod(validate_attribute_filters)
     )
 
-    _name = field_validator("name")(classmethod(valstr("name")))
+    @field_validator("name")
+    @classmethod
+    def _cant_set_none(cls, v):
+        return cant_set_none(v)
 
     @field_validator("zarr_dir")
     @classmethod